requirements.txt
torch>=2.0
fair-esm>=2.0.0
biopython>=1.80
scikit-learn>=1.3
scipy>=1.11
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
tqdm>=4.65
